fix: properly handle thumb aborts
