+++ packages/vexide-startup/Cargo.toml
@@ -15,6 +15,8 @@ authors = [
 ]
 
 [dependencies]
+arbitrary-int = { version = "2.1.1", optional = true }
+bitbybit = { version = "2.0.0", optional = true }
 bitflags = "2.4.2"
 vexide-core = { workspace = true }
 talc = { version = "4.3.1", optional = true }
@@ -48,7 +50,7 @@ workspace = true
 [features]
 allocator = ["dep:talc"]
 panic-hook = []
-abort-handler = []
+abort-handler = ["dep:bitbybit", "dep:arbitrary-int"]
 backtrace = ["dep:vex-libunwind", "dep:vex-libunwind-sys"]
 
 vex-sdk-jumptable = ["dep:vex-sdk-jumptable"]
